refactor(chat): isolate safe image sources
